Use common webpack config and add phosphor alias
