+++ core_package/webpack.config.js
@@ -3,6 +3,8 @@
 const data = require('./package.json');
 const Build = require('@jupyterlab/buildutils').Build;
 const webpack = require('webpack');
+const { merge } = require('webpack-merge');
+const baseConfig = require('../webpack.config.base');
 const { ModuleFederationPlugin } = webpack.container;
 const path = require('path');
 
@@ -16,51 +18,7 @@ const extras = Build.ensureAssets({
   output: './build'
 });
 
-const rules = [
-  { test: /\.css$/, use: ['style-loader', 'css-loader'] },
-  { test: /\.html$/, use: 'file-loader' },
-  { test: /\.md$/, use: 'raw-loader' },
-  { test: /\.(jpg|png|gif)$/, use: 'file-loader' },
-  { test: /\.js.map$/, use: 'file-loader' },
-  {
-    test: /\.woff2(\?v=\d+\.\d+\.\d+)?$/,
-    use: 'url-loader?limit=10000&mimetype=application/font-woff'
-  },
-  {
-    test: /\.woff(\?v=\d+\.\d+\.\d+)?$/,
-    use: 'url-loader?limit=10000&mimetype=application/font-woff'
-  },
-  {
-    test: /\.ttf(\?v=\d+\.\d+\.\d+)?$/,
-    use: 'url-loader?limit=10000&mimetype=application/octet-stream'
-  },
-  { test: /\.eot(\?v=\d+\.\d+\.\d+)?$/, use: 'file-loader' },
-  {
-    // In .css files, svg is loaded as a data URI.
-    test: /\.svg(\?v=\d+\.\d+\.\d+)?$/,
-    issuer: /\.css$/,
-    use: {
-      loader: 'svg-url-loader',
-      options: { encoding: 'none', limit: 10000 }
-    }
-  },
-  {
-    // In .ts and .tsx files (both of which compile to .js), svg files
-    // must be loaded as a raw string instead of data URIs.
-    test: /\.svg(\?v=\d+\.\d+\.\d+)?$/,
-    issuer: /\.js$/,
-    use: {
-      loader: 'raw-loader'
-    }
-  }
-];
-
-// TODO: make this configurable
-const options = {
-  devtool: 'source-map',
-  bail: true,
-  mode: 'development'
-};
+// TODO: make options configurable
 
 const singletons = {};
 
@@ -69,7 +27,7 @@ data.jupyterlab.singletonPackages.forEach(element => {
 });
 
 module.exports = [
-  {
+  merge(baseConfig, {
     entry: './index.js',
     output: {
       path: path.resolve(__dirname, 'build'),
@@ -81,10 +39,6 @@ module.exports = [
       // TODO: make this part of config so it can be different for other apps.
       publicPath: 'static/example/'
     },
-    stats: 'verbose',
-    ...options,
-    module: { rules },
-    resolve: { alias: { "url": false, "buffer": false } },
     plugins: [
       new ModuleFederationPlugin({
         library: {
@@ -96,11 +50,7 @@ module.exports = [
           ...data.dependencies,
           ...singletons
         }
-      }),
-      new webpack.DefinePlugin({
-        'process.env': '{}',
-        process: {}
       })
     ]
-  }
+  })
 ].concat(extras);
