+++ package.json
@@ -13,7 +13,8 @@
   "devDependencies": {
     "rimraf": "^3.0.2",
     "webpack": "5.0.0-beta.22",
-    "webpack-cli": "^3.3.10"
+    "webpack-cli": "^3.3.10",
+    "webpack-merge": "^5.0.9"
   },
   "workspaces": [
     "*"
