+++ webpack.config.ext.js
@@ -3,54 +3,11 @@
 
 const Build = require('@jupyterlab/buildutils').Build;
 const webpack = require('webpack');
+const { merge } = require('webpack-merge');
+const baseConfig = require('./webpack.config.base');
 const { ModuleFederationPlugin } = webpack.container;
 const path = require('path');
 const fs = require('fs');
-
-const rules = [
-  { test: /\.css$/, use: ['style-loader', 'css-loader'] },
-  { test: /\.html$/, use: 'file-loader' },
-  { test: /\.md$/, use: 'raw-loader' },
-  { test: /\.(jpg|png|gif)$/, use: 'file-loader' },
-  { test: /\.js.map$/, use: 'file-loader' },
-  {
-    test: /\.woff2(\?v=\d+\.\d+\.\d+)?$/,
-    use: 'url-loader?limit=10000&mimetype=application/font-woff'
-  },
-  {
-    test: /\.woff(\?v=\d+\.\d+\.\d+)?$/,
-    use: 'url-loader?limit=10000&mimetype=application/font-woff'
-  },
-  {
-    test: /\.ttf(\?v=\d+\.\d+\.\d+)?$/,
-    use: 'url-loader?limit=10000&mimetype=application/octet-stream'
-  },
-  { test: /\.eot(\?v=\d+\.\d+\.\d+)?$/, use: 'file-loader' },
-  {
-    // In .css files, svg is loaded as a data URI.
-    test: /\.svg(\?v=\d+\.\d+\.\d+)?$/,
-    issuer: /\.css$/,
-    use: {
-      loader: 'svg-url-loader',
-      options: { encoding: 'none', limit: 10000 }
-    }
-  },
-  {
-    // In .ts and .tsx files (both of which compile to .js), svg files
-    // must be loaded as a raw string instead of data URIs.
-    test: /\.svg(\?v=\d+\.\d+\.\d+)?$/,
-    issuer: /\.js$/,
-    use: {
-      loader: 'raw-loader'
-    }
-  }
-];
-
-const options = {
-  devtool: 'source-map',
-  bail: true,
-  mode: 'development'
-};
 
 const packagePath = process.env.PACKAGE_PATH;
 let outputPath = process.env.OUTPUT_PATH;
@@ -138,16 +95,13 @@ const bootstrap = 'import("' + exposes['./extension'] + '");'
 fs.writeFileSync(entryPoint, bootstrap);
 
 module.exports = [
-  {
+  merge(baseConfig, {
     entry: entryPoint,
     output: {
       filename: 'extension.js',
       path: outputPath,
       publicPath: `example/labextensions/${data.name}/`,
     },
-    ...options,
-    module: { rules },
-    resolve: { alias: { url: false, buffer: false } },
     plugins: [
       new ModuleFederationPlugin({
         name: data.name,
@@ -158,13 +112,9 @@ module.exports = [
         filename: 'remoteEntry.js',
         exposes,
         shared,
-      }),
-      new webpack.DefinePlugin({
-        'process.env': '{}',
-        process: {}
       })
     ]
-  }
+  })
 ].concat(extras);
 
 const logPath = path.join(outputPath, 'build_log.json');
